+++ src/test/kotlin/com/universal/hly/TestKotlin.kt
@@ -1,0 +1,11 @@
+package com.universal.hly
+
+import org.junit.Test
+
+class TestKotlin {
+    @Test
+    fun testString() {
+        val i = "100".toInt()
+        println(i)
+    }
+}